static analyzer error fixed

+++ rcsLib/rcsCmd/rcsCmd.cpp
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <string.h>
-#include "extra/ortsTypes/ortsTypes.h"
+#include <extra/ortsTypes/ortsTypes.h>
 #include "rcsCmd.h"
-#include "math/crc/crc16.h"
+#include <math/crc/crc16.h>
 
 rcsCmd::rcsCmd(BYTE serviceId, BYTE funcId)
 {
@@ -37,7 +37,7 @@ void rcsCmd::dbgPrint()
   printf(", Параметры [");
   if (cmd->func_paramsLength==0) printf (" ОТСУТСТВУЮТ ");
   else {
-	for (DWORD i=0; i<cmd->func_paramsLength; i++){
+	for (WORD i=0; i<cmd->func_paramsLength; i++){
 	    printf("%.2X ",*((BYTE*)cmd->func_params+i));
 	}
   }
@@ -66,7 +66,7 @@ errType rcsCmd::encode(const BYTE* data)
 	errType result=err_not_init;
 	
 	cmd->func_id=data[0];
-	cmd->func_paramsLength=*(DWORD*)(data+sizeof(cmd->func_id));
+	cmd->func_paramsLength=*(WORD*)(data+sizeof(cmd->func_id));
 	
 	if (cmd->func_paramsLength>0) {
 	    cmd->func_params=new BYTE[cmd->func_paramsLength];                                                            
@@ -186,7 +186,7 @@ errType rcsCmd::pushParam(OrtsType type, const void* param)
 	return result;
 }
 
-errType rcsCmd::encode(BYTE func_num, DWORD par_length, const void* data)
+errType rcsCmd::encode(BYTE func_num, WORD par_length, const void* data)
 {
 	errType result=err_not_init;
 	
@@ -225,7 +225,7 @@ DWORD rcsCmd::getCmdLength()
      return cmd->func_paramsLength+sizeof(cmd->func_id)+sizeof(cmd->crc16_signature)+sizeof(cmd->func_paramsLength);
 }
   
-const void* rcsCmd::get_func_paramsPtr(DWORD offset)
+const void* rcsCmd::get_func_paramsPtr(WORD offset)
 {
     return ((BYTE*)cmd->func_params+offset);
 }
